+++ tiledb/group.py
@@ -389,6 +389,10 @@ class Group(CtxMixin, lt.Group):
         return self._has_member(member)
 
     def __repr__(self):
+        # use safe repr if pybind11 constructor failed
+        if self._ctx is None:
+            return object.__repr__(self)
+
         return self._dump(True)
 
     def __enter__(self):